Create asst9
Question 1:

In this problem, we are interested in re-arranging the queue elements such that all the even integers are placed before the odd integers. Hence, you need to implement a method reArrange that will take a queue of integers as argument and will modify it accordingly. 
You are not allowed to use Queue methods other than: remove, peek, add, isEmpty and size. 
Your main method should: 
-	Prompt the user to enter N, the number of integers in the queue.
-	Prompt the user to enter the N integers. 
-	Invoke the method Rearrange.

Question 2:

Write a program to check if a given string is a palindrome using a stack. A palindrome is a string that reads the same forwards and backward, ignoring spaces, punctuation, and capitalization.